Split code in separate python modules
